Abstracting booking cancellation
